% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dh-LCM.R
\name{DhLCM}
\alias{DhLCM}
\title{Degree-heterogeneous latent class model estimation}
\usage{
DhLCM(
  R,
  K,
  spectral = "heteroPCA",
  norm = "L2",
  dist = "Bernoulli",
  T0 = 20,
  nstart = 10
)
}
\arguments{
\item{R}{Numeric matrix.}

\item{K}{Positive integer. The number of top eigen vectors to be extracted.}

\item{spectral}{Character. One of \code{"heteroPCA"} and \code{"SVD"}.
Specifies the method to be used to obtain the top \code{K} eigenvectors/left singular vectors.
\code{"heteroPCA"} implements the heteroPCA method.
\code{"SVD"} performs ordinary singular vector decomposition.}

\item{norm}{Character. One of \code{"L2"}, \code{"L1"}, \code{"SCORE"}, and \code{"NA"}.
Specifies the method to be used for normalization on the eigenvectors/left singular vectors.
\code{"L2"} performs L2 normalization.
\code{"L1"} performs L1 normalization.
\code{"SCORE"} performs SCORE normalization.
\code{"NA"} does not perform normalization.}

\item{dist}{Character. One of \code{"Bernoulli"}, \code{"Binomial"}, and \code{"Poisson"}.
Specifies the distribution of the ground truth model.
\code{"Bernoulli"} assumes the Bernoulli distribution.
\code{"Binomial"} assumes the Binomial distribution.
\code{"Poisson"} assumes the Poisson distribution.}

\item{T0}{Positive integer. The number of iterations for heteroPCA.}

\item{nstart}{Positive integer. The number of initial starts in the kmeans function.}
}
\value{
Named list. The list is made of:
\itemize{
\item \code{T_hat} --- Numeric matrix. Estimation of the \eqn{\Theta}{Theta} matrix.
\item \code{S_hat} --- Numeric vector. Clustered membership for each subject.
\item \code{Z_hat} --- Numeric matrix. Clustered membership for each subject in binary matrix form.
\item \code{sigma2_hat} --- Numeric vector (>=0). Asymptotic variance for each element of \code{T_hat}.
}
}
\description{
This function performs k-means clustering on the top \code{K} eigenvectors returned by heteroPCA, and estimates the DhLCM model parameters
}
