% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dh-LCM.R
\name{H_mat}
\alias{H_mat}
\title{Diagonal deletion}
\usage{
H_mat(X)
}
\arguments{
\item{X}{Numeric matrix}
}
\value{
A matrix of with diagonals set to 0
}
\description{
This function takes in a matrix, and returns the diagonal-deleted matrix
}
